additional tweaks

+++ spark_script.py
@@ -71,7 +71,7 @@ for Set in FilteredSets:
         if counter == 0:
             moments[counter] = 0
         else:
-            moments[counter] = moments[counter - 1] + (current_data - last_data) * 0.75
+            moments[counter] = moments[counter - 1] + double(current_data - last_data) * 0.75
         last_data = current_data
         counter = counter + 1
     post = {"FromEpoch": fromEpoch,
